chore: Start breaking down WPF Skia host to separate app hosting and window hosting
