+++ src/Uno.UI.Runtime.Skia.Gtk/GtkHost.cs
@@ -42,7 +42,7 @@ using Size = Windows.Foundation.Size;
 
 namespace Uno.UI.Runtime.Skia
 {
-	public partial class GtkHost : ISkiaHost
+	public partial class GtkHost : ISkiaApplicationHost
 	{
 		private const int UnoThemePriority = 800;
