+++ src/Uno.UI.Runtime.Skia.Wpf/Hosting/IWpfWindowHost.cs
@@ -1,0 +1,12 @@
+﻿#nullable enable
+
+using Uno;
+using Windows.UI.Xaml;
+using WpfCanvas = System.Windows.Controls.Canvas;
+
+namespace Uno.UI.Runtime.Skia.Wpf.Hosting;
+
+internal interface IWpfWindowHost : IWpfXamlRootHost
+{
+	WpfCanvas? NativeOverlayLayer { get; }
+}
