+++ src/Uno.UI.Runtime.Skia.Wpf/Hosting/IWpfXamlRootHost.cs
@@ -1,0 +1,14 @@
+﻿#nullable enable
+
+using Uno;
+using Windows.UI.Xaml;
+using WpfCanvas = System.Windows.Controls.Canvas;
+
+namespace Uno.UI.Runtime.Skia.Wpf.Hosting;
+
+internal interface IWpfXamlRootHost
+{
+	WpfCanvas? NativeOverlayLayer { get; }
+
+	XamlRoot? XamlRoot { get; }
+}
