+++ src/Uno.UI.Runtime.Skia.Wpf/Rendering/OpenGLWpfRenderer.cs
@@ -21,7 +21,7 @@ internal partial class OpenGLWpfRenderer : IWpfRenderer
 	private const GRSurfaceOrigin surfaceOrigin = GRSurfaceOrigin.TopLeft;
 
 	private readonly WpfControl _hostControl;
-	private readonly IWpfHost _host;
+	private readonly IWpfApplicationHost _host;
 	private DisplayInformation? _displayInformation;
 	private nint _hwnd;
 	private nint _hdc;
@@ -31,7 +31,7 @@ internal partial class OpenGLWpfRenderer : IWpfRenderer
 	private GRBackendRenderTarget? _renderTarget;
 	private WriteableBitmap? _backBuffer;
 
-	public OpenGLWpfRenderer(IWpfHost host)
+	public OpenGLWpfRenderer(IWpfApplicationHost host)
 	{
 		_hostControl = host as WpfControl ?? throw new InvalidOperationException("Host should be a WPF control");
 		_host = host;
