+++ src/Uno.UI.Runtime.Skia.Wpf/Themes/Generic.xaml
@@ -1,10 +1,10 @@
 <ResourceDictionary
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:platform="clr-namespace:Uno.UI.Skia.Platform"
+    xmlns:platform="clr-namespace:Uno.UI.Skia.Wpf"
     xmlns:controls="clr-namespace:Uno.UI.Runtime.Skia.Wpf.Controls">
 
-    <Style TargetType="{x:Type platform:WpfHost}">
+    <Style TargetType="{x:Type platform:UnoWindow}">
         <Setter Property="Template">
             <Setter.Value>
 				<ControlTemplate TargetType="{x:Type platform:WpfHost}">
