+++ src/Uno.UI/UI/Xaml/ISkiaApplicationHost.cs
@@ -1,0 +1,5 @@
+﻿namespace Windows.UI.Xaml;
+
+internal interface ISkiaApplicationHost
+{
+}
