--- src/Uno.UI/UI/Xaml/ISkiaHost.cs
@@ -1,5 +1,0 @@
-﻿namespace Windows.UI.Xaml;
-
-internal interface ISkiaHost
-{
-}
